fixes due to commets for task 2
