+++ src/app/core/components/footer/footer.component.spec.ts
