+++ src/app/core/components/header/header.component.html
@@ -1,5 +1,5 @@
 <mat-toolbar color="primary" class="container">
-  <img src="assets/images/logo.png" alt="Logo" class="logo">
+  <img src="../../../../assets/images/logo.png" alt="Logo" class="logo">
   <span>VIDEO COURSE</span>
 
   <span class="spacer"></span>
