+++ src/app/core/models/course.ts
