+++ src/app/courses/components/course-item/course-item.component.css
