+++ src/app/courses/components/course-item/course-item.component.spec.ts
