+++ src/app/courses/components/course-item/course-item.component.ts
@@ -1,16 +1,17 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { ICourse } from '../../models/course';
+import { ICourse } from '../../../core/models/course';
 
 @Component({
   selector: 'app-course-item',
   templateUrl: './course-item.component.html',
   styleUrls: ['./course-item.component.css']
 })
+
 export class CourseItemComponent implements OnInit {
-  @Input() course: ICourse;
+  @Input() course!: ICourse;
   @Output() deletedCourseId = new EventEmitter<number>();
-  creationDate: Date;
-  formattedDate: string;
+  creationDate!: Date;
+  formattedDate!: string;
 
   ngOnInit() {
     this.creationDate = new Date(this.course.creationDate);
