+++ src/app/courses/components/course-list/course-list.component.css
