+++ src/app/courses/courses.module.ts
@@ -5,11 +5,11 @@ import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatToolbarModule } from '@angular/material/toolbar';
-import { CourseListComponent } from './course-list/course-list.component';
-import { CourseItemComponent } from './course-item/course-item.component';
-import { SectionComponent } from './section/section.component';
-import { SearchCourseComponent } from './search-course/search-course.component';
-import { AddCourseComponent } from './add-course/add-course.component';
+import { CourseListComponent } from './components/course-list/course-list.component';
+import { CourseItemComponent } from './components/course-item/course-item.component';
+import { SectionComponent } from './components/section/section.component';
+import { SearchCourseComponent } from './components/search-course/search-course.component';
+import { AddCourseComponent } from './components/add-course/add-course.component';
 
 @NgModule({
   declarations: [
