+++ tsconfig.json
@@ -7,7 +7,7 @@
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "noImplicitOverride": true,
-    "noPropertyAccessFromIndexSignature": false,
+    "noPropertyAccessFromIndexSignature": true,
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "sourceMap": true,
@@ -18,7 +18,6 @@
     "importHelpers": true,
     "target": "ES2022",
     "module": "ES2022",
-    "strictPropertyInitialization": false,
     "useDefineForClassFields": false,
     "lib": [
       "ES2022",
